# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CornerNote {
  id: string;
  number: number;
  name: string;
  gear: string;
  brakingPoint: string;
  lineNotes: string;
  myNotes: string;
}

export interface LapRecord {
  lap: number;
  time: string;
  s1: string;
  s2: string;
  s3: string;
  tires: string;
  penalty: string;
}

export interface SessionRecord {
  id: string;
  date: string;
  trackId: string;
  car: string;
  type: string;
  bestLap: string;
  avgLap: string;
  worstLap: string;
  s1: string;
  s2: string;
  s3: string;
  tires: string;
  fuelLoad: number;
  conditions: string;
  assists: string;
  rating: number;
  notes: string;
  isPB: boolean;
  penalty?: string | null;
  gameVersion?: string | null;
  platform?: string | null;
  inputDevice?: string | null;
  isPublic: boolean;
  sharedAt?: string | null;
  publicNote?: string | null;
  laps?: LapRecord[] | null;
  position?: string;
}

export interface CreateSessionRequest {
  id: string;
  date: string;
  trackId: string;
  car: string;
  type: string;
  bestLap: string;
  avgLap: string;
  worstLap: string;
  s1: string;
  s2: string;
  s3: string;
  tires: string;
  fuelLoad: number;
  conditions: string;
  assists: string;
  rating: number;
  notes: string;
  penalty?: string;
  gameVersion?: string;
  platform?: string;
  inputDevice?: string;
  laps?: LapRecord[];
  position?: string;
}

export interface ShareSessionRequest {
  publicNote?: string;
}

export interface ShareSessionResponse {
  isPublic: boolean;
  sharedAt?: string | null;
}

export interface CommunitySessionRecord {
  id: string;
  date: string;
  trackId: string;
  car: string;
  type: string;
  bestLap: string;
  avgLap: string;
  tires: string;
  conditions: string;
  penalty?: string | null;
  gameVersion?: string | null;
  platform?: string | null;
  inputDevice?: string | null;
  publicNote?: string | null;
  authorName: string;
  isOwn: boolean;
  sharedAt?: string | null;
  rating: number;
}

export interface SetupRecord {
  id: string;
  label: string;
  car: string;
  trackId: string;
  tag: string;
  date: string;
  frontWing: string;
  rearWing: string;
  frontARB: string;
  rearARB: string;
  frontRideHeight: string;
  rearRideHeight: string;
  frontSprings: string;
  rearSprings: string;
  brakeBias: string;
  brakePressure: string;
  onThrottle: string;
  offThrottle: string;
  notes: string;
  isPublic?: boolean;
  sharedAt?: string | null;
  gameVersion?: string | null;
}

export interface CommunitySetupRecord {
  id: string;
  label: string;
  car: string;
  trackId: string;
  tag: string;
  date: string;
  frontWing: string;
  rearWing: string;
  frontARB: string;
  rearARB: string;
  frontRideHeight: string;
  rearRideHeight: string;
  frontSprings: string;
  rearSprings: string;
  brakeBias: string;
  brakePressure: string;
  onThrottle: string;
  offThrottle: string;
  notes: string;
  authorName: string;
  isOwn: boolean;
  avgRating?: number | null;
  ratingCount: number;
  sharedAt?: string | null;
}

export interface ShareSetupResponse {
  isPublic: boolean;
  sharedAt?: string | null;
}

export interface RateSetupRequest {
  /**
     * @minimum 1
     * @maximum 5
     */
  stars: number;
}

export interface RateSetupResponse {
  avgRating?: number | null;
  ratingCount: number;
}

export interface CreateSetupRequest {
  id: string;
  label: string;
  car: string;
  trackId: string;
  tag?: string;
  date: string;
  frontWing: string;
  rearWing: string;
  frontARB: string;
  rearARB: string;
  frontRideHeight: string;
  rearRideHeight: string;
  frontSprings: string;
  rearSprings: string;
  brakeBias: string;
  brakePressure: string;
  onThrottle: string;
  offThrottle: string;
  notes: string;
  gameVersion?: string;
}

export interface TrackNotesRecord {
  id: string;
  trackId: string;
  corners: CornerNote[];
}

export interface UpsertTrackNotesRequest {
  id: string;
  corners: CornerNote[];
}

export interface HardwareRecord {
  id: string;
  label: string;
  peripheralType: string;
  brand: string;
  model: string;
  trackId: string;
  game: string;
  date: string;
  ffbStrength: string;
  maxForce: string;
  damper: string;
  friction: string;
  linearity: string;
  steeringRange: string;
  notes: string;
}

export interface CreateHardwareRequest {
  id: string;
  label: string;
  peripheralType: string;
  brand: string;
  model: string;
  trackId: string;
  game: string;
  date: string;
  ffbStrength: string;
  maxForce: string;
  damper: string;
  friction: string;
  linearity: string;
  steeringRange: string;
  notes: string;
}

/**
 * Unauthorized
 */
export type UnauthorizedResponse = ErrorResponse;

/**
 * Not found
 */
export type NotFoundResponse = ErrorResponse;

export type GetCommunitySessionsParams = {
sort?: string;
};

export type GetCommunitySetupsParams = {
trackId?: string;
car?: string;
tag?: string;
};