# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community/sessions:
    get:
      operationId: getCommunitySessions
      tags: [community]
      summary: Get all publicly shared sessions
      parameters:
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of community sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunitySessionRecord"
  /community/setups:
    get:
      operationId: getCommunitySetups
      tags: [community]
      summary: Get all publicly shared setups
      parameters:
        - name: trackId
          in: query
          required: false
          schema:
            type: string
        - name: car
          in: query
          required: false
          schema:
            type: string
        - name: tag
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of community setups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunitySetupRecord"
  /community/setups/{id}:
    get:
      operationId: getCommunitySetup
      tags: [community]
      summary: Get a single public community setup
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Community setup
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunitySetupRecord"
        "404":
          $ref: "#/components/responses/NotFound"
  /community/setups/{id}/rate:
    post:
      operationId: rateSetup
      tags: [community]
      summary: Rate a community setup (1-5 stars)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RateSetupRequest"
      responses:
        "200":
          description: Rating saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateSetupResponse"
        "400":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotFound"
        "404":
          $ref: "#/components/responses/NotFound"
  /community/setups/{id}/import:
    post:
      operationId: importSetup
      tags: [community]
      summary: Import a community setup into the caller's vault
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Setup imported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SetupRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CommunitySessionRecord:
      type: object
      properties:
        id:
          type: string
        date:
          type: string
        trackId:
          type: string
        car:
          type: string
        type:
          type: string
        bestLap:
          type: string
        avgLap:
          type: string
        tires:
          type: string
        conditions:
          type: string
        penalty:
          type: string
          nullable: true
        gameVersion:
          type: string
          nullable: true
        platform:
          type: string
          nullable: true
        inputDevice:
          type: string
          nullable: true
        publicNote:
          type: string
          nullable: true
        authorName:
          type: string
        isOwn:
          type: boolean
        sharedAt:
          type: string
          nullable: true
        rating:
          type: integer
      required:
        - id
        - date
        - trackId
        - car
        - type
        - bestLap
        - avgLap
        - tires
        - conditions
        - authorName
        - isOwn
        - rating
    SetupRecord:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        car:
          type: string
        trackId:
          type: string
        tag:
          type: string
        date:
          type: string
        frontWing:
          type: string
        rearWing:
          type: string
        frontARB:
          type: string
        rearARB:
          type: string
        frontRideHeight:
          type: string
        rearRideHeight:
          type: string
        frontSprings:
          type: string
        rearSprings:
          type: string
        brakeBias:
          type: string
        brakePressure:
          type: string
        onThrottle:
          type: string
        offThrottle:
          type: string
        notes:
          type: string
        isPublic:
          type: boolean
        sharedAt:
          type: string
          nullable: true
        gameVersion:
          type: string
          nullable: true
      required:
        - id
        - label
        - car
        - trackId
        - tag
        - date
        - frontWing
        - rearWing
        - frontARB
        - rearARB
        - frontRideHeight
        - rearRideHeight
        - frontSprings
        - rearSprings
        - brakeBias
        - brakePressure
        - onThrottle
        - offThrottle
        - notes
    CommunitySetupRecord:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        car:
          type: string
        trackId:
          type: string
        tag:
          type: string
        date:
          type: string
        frontWing:
          type: string
        rearWing:
          type: string
        frontARB:
          type: string
        rearARB:
          type: string
        frontRideHeight:
          type: string
        rearRideHeight:
          type: string
        frontSprings:
          type: string
        rearSprings:
          type: string
        brakeBias:
          type: string
        brakePressure:
          type: string
        onThrottle:
          type: string
        offThrottle:
          type: string
        notes:
          type: string
        authorName:
          type: string
        isOwn:
          type: boolean
        avgRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
        sharedAt:
          type: string
          nullable: true
      required:
        - id
        - label
        - car
        - trackId
        - tag
        - date
        - frontWing
        - rearWing
        - frontARB
        - rearARB
        - frontRideHeight
        - rearRideHeight
        - frontSprings
        - rearSprings
        - brakeBias
        - brakePressure
        - onThrottle
        - offThrottle
        - notes
        - authorName
        - isOwn
        - ratingCount
...
    RateSetupRequest:
      type: object
      properties:
        stars:
          type: integer
          minimum: 1
          maximum: 5
      required:
        - stars
    RateSetupResponse:
      type: object
      properties:
        avgRating:
          type: number
          nullable: true
        ratingCount:
          type: integer
      required:
        - ratingCount
...